requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
pandas
numpy
scikit-learn
openpyxl
tqdm
asyncio
db-dtype
